add 1d convolution

+++ include/wavelet.h
@@ -1,6 +1,6 @@
 #ifndef WAVELET_H
 #define WAVELET_H
 
-void ricker(double *arr, int Nt, double dt, int fmax);
+void ricker(float *arr, int Nt, double dt, int fmax);
 
 #endif // !WAVELET_H
